Moved EventFormatProvider and ExtensionsParser into provider Renamed ExtensionsParser in ExtensionProvider
