+++ http/restful-ws-integration-tests/restful-ws-jersey/src/test/java/io/cloudevents/http/restful/ws/jersey/TestJersey.java
@@ -18,8 +18,8 @@
 package io.cloudevents.http.restful.ws.jersey;
 
 import com.github.hanleyt.JerseyExtension;
-import io.cloudevents.core.format.EventFormatProvider;
 import io.cloudevents.core.mock.CSVFormat;
+import io.cloudevents.core.provider.EventFormatProvider;
 import io.cloudevents.http.restful.ws.BaseTest;
 import io.cloudevents.http.restful.ws.CloudEventsProvider;
 import io.cloudevents.http.restful.ws.TestResource;
