+++ http/restful-ws/src/main/java/io/cloudevents/http/restful/ws/CloudEventsProvider.java
@@ -19,9 +19,9 @@ package io.cloudevents.http.restful.ws;
 
 import io.cloudevents.CloudEvent;
 import io.cloudevents.core.format.EventFormat;
-import io.cloudevents.core.format.EventFormatProvider;
 import io.cloudevents.core.message.Message;
 import io.cloudevents.core.message.MessageVisitor;
+import io.cloudevents.core.provider.EventFormatProvider;
 import io.cloudevents.http.restful.ws.impl.RestfulWSClientMessageVisitor;
 import io.cloudevents.http.restful.ws.impl.RestfulWSMessageFactory;
 import io.cloudevents.http.restful.ws.impl.RestfulWSMessageVisitor;
