+++ kafka/src/main/java/io/cloudevents/kafka/CloudEventSerializer.java
@@ -19,9 +19,9 @@ package io.cloudevents.kafka;
 
 import io.cloudevents.CloudEvent;
 import io.cloudevents.core.format.EventFormat;
-import io.cloudevents.core.format.EventFormatProvider;
 import io.cloudevents.core.message.Encoding;
 import io.cloudevents.core.message.Message;
+import io.cloudevents.core.provider.EventFormatProvider;
 import io.cloudevents.kafka.impl.KafkaSerializerMessageVisitorImpl;
 import org.apache.kafka.common.header.Headers;
 import org.apache.kafka.common.serialization.Serializer;
