+++ kafka/src/test/java/io/cloudevents/kafka/CloudEventSerializerTest.java
@@ -18,10 +18,10 @@
 package io.cloudevents.kafka;
 
 import io.cloudevents.CloudEvent;
-import io.cloudevents.core.format.EventFormatProvider;
 import io.cloudevents.core.message.Encoding;
 import io.cloudevents.core.message.Message;
 import io.cloudevents.core.mock.CSVFormat;
+import io.cloudevents.core.provider.EventFormatProvider;
 import io.cloudevents.core.test.Data;
 import org.apache.kafka.common.header.Headers;
 import org.apache.kafka.common.header.internals.RecordHeaders;
